refactor(harvester): move methods for additional metadata handling to separate file
